requirements.txt
aiohttp==2.3.10
alembic==0.9.7
async-timeout==2.0.0
asyncio==3.4.3
beautifulsoup4==4.6.0
bleach==2.1.3
certifi==2018.1.18
chardet==3.0.4
choix==0.3.1
cssselect==1.0.1
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
fake-useragent==0.1.10
greenlet==0.4.13
html2text==2017.10.4
html5lib==1.0.1
idna==2.6
idna-ssl==1.0.0
ImageHash==4.0
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lxml==3.8.0
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.2.0
mistune==0.8.3
msgpack==0.5.4
multidict==4.1.0
nbconvert==5.3.1
nbformat==4.4.0
neovim==0.2.1
notebook==5.4.0
numpy==1.14.0
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
psutil==5.4.3
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
PySocks==1.6.7
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2018.3
PyWavelets==0.5.2
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
selenium==3.8.1
Send2Trash==1.5.0
SerpScrap==0.10.1
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.2
terminado==0.8.1
testpath==0.3.1
tornado==5.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.4
yarl==1.1.1
tensorflow==1.6.0
imageio==2.2.0
